Add spaces, special characters and upcase letters
First, the code restores the spaces and special characters by checking
whether any of these elements appear in the original string. When it
finds one, it uses .insert to place the character back into the correct
index.

Finally, to restore uppercase letters, the code checks each character’s
corresponding position in the original string. Using map, if the
original character was uppercase, it applies .upcase to ensure the
transformed string matches that casing.